Abstract to SocketAddress
This updates some methods for cross platform use as well as unifying
_unix files.
