+++ runtime/v2/shim/shim_darwin.go
@@ -18,25 +18,9 @@
 
 package shim
 
-import (
-	"os"
-	"os/signal"
-
-	"github.com/containerd/ttrpc"
-)
-
-// setupSignals creates a new signal handler for all signals and sets the shim as a
-// sub-reaper so that the container processes are reparented
-func setupSignals() (chan os.Signal, error) {
-	signals := make(chan os.Signal, 2048)
-	signal.Notify(signals)
-	return signals, nil
-}
+import "github.com/containerd/ttrpc"
 
 func newServer() (*ttrpc.Server, error) {
-	// for darwin, we omit the socket credentials because these syscalls are
-	// slightly different. since we don't have darwin support yet, this can be
-	// implemented later and the build can continue without issue.
 	return ttrpc.NewServer()
 }
