+++ runtime/v2/shim/shim_linux.go
@@ -17,21 +17,9 @@
 package shim
 
 import (
-	"os"
-	"os/signal"
-
 	"github.com/containerd/containerd/sys"
 	"github.com/containerd/ttrpc"
-	"golang.org/x/sys/unix"
 )
-
-// setupSignals creates a new signal handler for all signals and sets the shim as a
-// sub-reaper so that the container processes are reparented
-func setupSignals() (chan os.Signal, error) {
-	signals := make(chan os.Signal, 32)
-	signal.Notify(signals, unix.SIGTERM, unix.SIGINT, unix.SIGCHLD, unix.SIGPIPE)
-	return signals, nil
-}
 
 func newServer() (*ttrpc.Server, error) {
 	return ttrpc.NewServer(ttrpc.WithServerHandshaker(ttrpc.UnixSocketRequireSameUser()))
